fix: move localRenderer out of test dir
test dir isnt included in docker image, so exporting it in index.js is breaking the image, but working fine when installed via npm. Moving it out in a new lib dir so it is included in the both the image and npm package
