+++ lib/LocalMustacheTemplateController.js
@@ -1,5 +1,5 @@
 const { MockKubeResourceMeta } = require('@razee/razeedeploy-core');
-const MustacheTemplateController = require('../../src/MustacheTemplateController');
+const MustacheTemplateController = require('../src/MustacheTemplateController');
 
 module.exports = class LocalMustacheTemplateController extends MustacheTemplateController {
   constructor(params) {
