+++ lib/LocalRenderer.js
